+++ routes/todo.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter, HTTPException, Depends
 from typing import Annotated
 
 from schemas.todo import TodoCreate, TodoUpdate
+from schemas.user import UserSchema
 from services import todo as todo_svc
 from dependencies.db import get_db
 from dependencies.user import get_current_verified_user
@@ -14,42 +15,73 @@ router = APIRouter(
 )
 
 
-@router.get("/{todo_id}")
-def get_todo(todo_id: str, db=Depends(get_db)):
+@router.get("{organization_id}/{todo_id}")
+def get_todo(
+    organization_id: str,
+    todo_id: str,
+    user: Annotated[UserSchema, Depends(get_current_verified_user)],
+    db=Depends(get_db),
+):
     try:
+        check_permission(db, user, organization_id, "view_task")
         return todo_svc.get_todo(db, todo_id)
     except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
 
 
-@router.get("/{folder_id}")
-def get_todos(folder_id: str, db=Depends(get_db)):
+@router.get("{organization_id}/{folder_id}")
+def get_todos(
+    organization_id: str,
+    folder_id: str,
+    user: Annotated[UserSchema, Depends(get_current_verified_user)],
+    db=Depends(get_db),
+):
     try:
+        check_permission(db, user, organization_id, "list_task")
         return todo_svc.get_todos(db, folder_id)
     except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
 
 
-@router.post("")
-def create_todo(todo: TodoCreate, db=Depends(get_db)):
+@router.post("{organization_id}")
+def create_todo(
+    organization_id: str,
+    todo: TodoCreate,
+    user: Annotated[UserSchema, Depends(get_current_verified_user)],
+    db=Depends(get_db),
+):
     try:
+        check_permission(db, user, organization_id, "create_task")
         return todo_svc.create_todo(db, todo)
     except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
 
 
-@router.patch("/{todo_id}")
-def update_todo(todo_id: str, todo: TodoUpdate, db=Depends(get_db)):
+@router.patch("{organization_id}/{todo_id}")
+def update_todo(
+    organization_id: str,
+    todo_id: str,
+    todo: TodoUpdate,
+    user: Annotated[UserSchema, Depends(get_current_verified_user)],
+    db=Depends(get_db),
+):
     try:
+        check_permission(db, user, organization_id, "update_task")
         todo_svc.get_todo(db, todo_id)
         return todo_svc.update_todo(db, todo)
     except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
 
 
-@router.delete("/{todo_id}")
-def delete_todo(todo_id: str, db=Depends(get_db)):
+@router.delete("{organization_id}/{todo_id}")
+def delete_todo(
+    organization_id: str,
+    todo_id: str,
+    user: Annotated[UserSchema, Depends(get_current_verified_user)],
+    db=Depends(get_db),
+):
     try:
+        check_permission(db, user, organization_id, "delete_task")
         return todo_svc.delete_todo(db, todo_id)
     except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
